requirements.txt
requests==2.31.0
Pillow==10.3.0
python-dotenv==1.0.0
google-api-python-client==2.131.0
google-auth-oauthlib==1.2.0
anthropic==0.25.0